roles and other stuff from fork

+++ Domain/Grant.cs
@@ -2,6 +2,6 @@
 {
     public class Grant : Event
     {
-        public decimal GrantAmount { get; set; }
+        public double GrantAmount { get; set; }
     }
 }
